Feb 8 Changes

+++ S2/p9.cpp
@@ -5,7 +5,7 @@ int main() {
     char character;
     cout << "Enter a character";
     cin >> character;
-    int ascii = (int)character;
+    int ascii = character;
     cout << "The ascii value of your character is: " << ascii;
